+++ src/command/mod.rs
@@ -15,3 +15,4 @@ pub mod ps;
 pub mod push;
 pub mod standalone;
 pub mod topology;
+pub mod versions;